+++ changelog.md
@@ -184,6 +184,7 @@ Proof-of-concept version.
 ## 1.2.9
 
 * MSVC fix: added the `/DEBUG` flag during linking to allow actual debugging in debug mode
+* When a compile toolchain element is missing, cook will now report what source files are actually requiring this language support.
 
 ## Next